QA: Unify dmn setup code into test_framework
