+++ test/functional/test_framework/test_framework.py
@@ -1337,7 +1337,104 @@ class PivxTestFramework():
         assert_equal(pl["operatorPubKey"], dmn.operator_pk)
         assert_equal(pl["payoutAddress"], dmn.payee)
 
-# LLMQs
+# ------------------------------------------------------
+
+class SkipTest(Exception):
+    """This exception is raised to skip a test"""
+    def __init__(self, message):
+        self.message = message
+
+
+'''
+PivxTestFramework extensions
+'''
+class PivxDMNTestFramework(PivxTestFramework):
+
+    def set_base_test_params(self):
+        # 1 miner, 1 controller, 6 remote mns
+        self.num_nodes = 8
+        self.minerPos = 0
+        self.controllerPos = 1
+        self.setup_clean_chain = True
+
+    def add_new_dmn(self, strType, op_keys=None, from_out=None):
+        self.mns.append(self.register_new_dmn(2 + len(self.mns),
+                                             self.minerPos,
+                                             self.controllerPos,
+                                             strType,
+                                             outpoint=from_out,
+                                             op_blskeys=op_keys))
+
+    def check_mn_list(self):
+        for i in range(self.num_nodes):
+            self.check_mn_list_on_node(i, self.mns)
+        self.log.info("Deterministic list contains %d masternodes for all peers." % len(self.mns))
+
+    def check_mn_enabled_count(self, enabled, total):
+        for node in self.nodes:
+            node_count = node.getmasternodecount()
+            assert_equal(node_count['enabled'], enabled)
+            assert_equal(node_count['total'], total)
+
+    def wait_until_mnsync_completed(self):
+        SYNC_FINISHED = [999] * self.num_nodes
+        synced = [-1] * self.num_nodes
+        timeout = time.time() + 120
+        while synced != SYNC_FINISHED and time.time() < timeout:
+            synced = [node.mnsync("status")["RequestedMasternodeAssets"]
+                      for node in self.nodes]
+            if synced != SYNC_FINISHED:
+                time.sleep(5)
+        if synced != SYNC_FINISHED:
+            raise AssertionError("Unable to complete mnsync: %s" % str(synced))
+
+    def setup_test(self):
+        self.mns = []
+        self.disable_mocktime()
+        connect_nodes_clique(self.nodes)
+
+        # Enforce mn payments and reject legacy mns at block 131
+        self.activate_spork(0, "SPORK_8_MASTERNODE_PAYMENT_ENFORCEMENT")
+        assert_equal("success", self.set_spork(self.minerPos, "SPORK_21_LEGACY_MNS_MAX_HEIGHT", 130))
+        time.sleep(1)
+        assert_equal([130] * self.num_nodes, [self.get_spork(x, "SPORK_21_LEGACY_MNS_MAX_HEIGHT")
+                                              for x in range(self.num_nodes)])
+
+        # Mine 130 blocks
+        self.log.info("Mining...")
+        self.nodes[self.minerPos].generate(10)
+        self.sync_blocks()
+        self.wait_until_mnsync_completed()
+        self.nodes[self.minerPos].generate(120)
+        self.sync_blocks()
+        self.assert_equal_for_all(130, "getblockcount")
+
+        # enabled/total masternodes: 0/0
+        self.check_mn_enabled_count(0, 0)
+
+        # Create 6 DMNs and init the remote nodes
+        self.log.info("Initializing masternodes...")
+        for _ in range(2):
+            self.add_new_dmn("internal")
+            self.add_new_dmn("external")
+            self.add_new_dmn("fund")
+        assert_equal(len(self.mns), 6)
+        for mn in self.mns:
+            self.nodes[mn.idx].initmasternode(mn.operator_sk, "", True)
+            time.sleep(1)
+        self.nodes[self.minerPos].generate(1)
+        self.sync_blocks()
+
+        # enabled/total masternodes: 6/6
+        self.check_mn_enabled_count(6, 6)
+        self.check_mn_list()
+
+        # Check status from remote nodes
+        assert_equal([self.nodes[idx].getmasternodestatus()['status'] for idx in range(2, self.num_nodes)],
+                     ["Ready"] * (self.num_nodes - 2))
+        self.log.info("All masternodes ready.")
+
+    # LLMQ functions
 
     def get_quorum_connections(self, node, members):
         conn = []
@@ -1443,7 +1540,7 @@ class PivxTestFramework():
     """
     def mine_quorum(self, invalidate_func=None, invalidated_idx=None, skip_bad_member_sync=False):
         nodes_to_sync = self.nodes.copy()
-        if invalidated_idx:
+        if invalidated_idx is not None:
             assert invalidate_func is None
             if skip_bad_member_sync:
                 nodes_to_sync.pop(invalidated_idx)
@@ -1501,18 +1598,6 @@ class PivxTestFramework():
         self.log.info("Final commitment correctly mined on chain")
         return qfc, bad_member
 
-
-# ------------------------------------------------------
-
-class SkipTest(Exception):
-    """This exception is raised to skip a test"""
-    def __init__(self, message):
-        self.message = message
-
-
-'''
-PivxTestFramework extensions
-'''
 # !TODO: remove after obsoleting legacy system
 class PivxTier2TestFramework(PivxTestFramework):
